add LSB operation
